+++ src/operator/tensor/broadcast_reduce_sum_value.cc
@@ -27,35 +27,6 @@
 namespace mxnet {
 namespace op {
 
-NumpyReduceAxesParam ConvertReduceNDParamsToNumpy(const ReduceAxesParam& original_param,
-                                                  const NDArray& input,
-                                                  const NDArray& output) {
-  NumpyReduceAxesParam numpy_param;
-  numpy_param.axis = dmlc::optional<mxnet::Tuple<int>>();
-  if (original_param.axis.has_value()) {
-    mxnet::Tuple<int> axes(original_param.axis.value().begin(), original_param.axis.value().end());
-    std::sort(axes.begin(), axes.end());
-
-    if (original_param.exclude) {
-      const size_t in_ndim = input.shape().ndim();
-      mxnet::Tuple<int> inverted_axes(in_ndim - axes.ndim(), -1);
-      for (int i = 0, j = 0; i < input.shape().ndim(); i++) {
-        if (j >= axes.ndim() || i != axes[j]) {
-          inverted_axes[i - j] = i;
-        } else {
-          j++;
-        }
-      }
-      numpy_param.axis = inverted_axes;
-    } else {
-      numpy_param.axis = axes;
-    }
-  }
-  numpy_param.keepdims = original_param.keepdims;
-  numpy_param.dtype    = dmlc::optional<int>(output.dtype());
-  return numpy_param;
-}
-
 MXNET_OPERATOR_REGISTER_REDUCE(sum)
 MXNET_ADD_SPARSE_OP_ALIAS(sum)
     .add_alias("sum_axis")
